Add shortcuts to resize window and command to reindent

+++ config.sh
@@ -1,2 +1,16 @@
-rm ~/.vim/vimrc
+#!usr/bin/sh
+
+if [! -d "~/.vim"]; then
+    mkdir ~/.vim
+fi
+
+if [! -f "~/.vim/vimrc"]; then
+    rm ~/.vim/vimrc
+fi
+
 cp vimrc ~/.vim/
+
+if [! -d "~/.vim/colors"]; then
+    cp -r ./colors ~/.vim
+fi
+
